refactor: move BookingCard component to Cards directory and remove unused file
